% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{treedataverse_deps}
\alias{treedataverse_deps}
\title{List all treedataverse dependencies}
\usage{
treedataverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
treedataverse packages.}
}
\description{
List all treedataverse dependencies
}
