% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{treedataverse_packages}
\alias{treedataverse_packages}
\title{List all packages in the treedataverse}
\usage{
treedataverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include treedataverse in the list?}
}
\description{
List all packages in the treedataverse
}
\examples{
treedataverse_packages()
}
