% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{treedataverse_update}
\alias{treedataverse_update}
\title{Update treedataverse packages}
\usage{
treedataverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
treedataverse packages.}
}
\description{
This will check to see if all treedataverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
treedataverse_update()
}
}
